Clean up code

+++ source/d3d11/d3d11_device.hpp
@@ -104,7 +104,7 @@ struct DECLSPEC_UUID("72299288-2C68-4AD8-945D-2BFB5AA9C609") D3D11Device final :
 	using device_impl::_orig;
 
 	LONG _ref = 1;
-	unsigned int _interface_version = 0;
+	unsigned short _interface_version = 0;
 	D3D11On12Device *_d3d11on12_device = nullptr;
 	D3D11DeviceContext *_immediate_context = nullptr;
 };
