+++ source/dxgi/dxgi_device.hpp
@@ -44,5 +44,5 @@ struct DECLSPEC_UUID("CB285C3B-3677-4332-98C7-D6339B9782B1") DXGIDevice : IDXGID
 	bool check_and_upgrade_interface(REFIID riid);
 
 	IDXGIDevice1 *_orig;
-	unsigned int _interface_version;
+	unsigned short _interface_version;
 };
