+++ source/dxgi/dxgi_swapchain.hpp
@@ -91,7 +91,7 @@ struct DECLSPEC_UUID("1F445F9F-9887-4C4C-9055-4E3BADAFCCA8") DXGISwapChain final
 
 	LONG _ref = 1;
 	IDXGISwapChain *_orig;
-	unsigned int _interface_version;
+	unsigned short _interface_version;
 	IUnknown *const _direct3d_device;
 	IUnknown *const _direct3d_command_queue;
 	const unsigned int _direct3d_version;
